Combine the AIAutoIdlePlugin into AIAutomaticStatus so we can have a finer level of control on when we unset an automatic idle event.
Adds a confirmation window (option defaulting to NO in the Status Preferences) for the return from an automatic change (idle report, idle status, FUS, SS). When a user requests the status not be reverted, and an idle time is present, every account's status (if it wasn't forcing an idle time before) is duplicated with a forced idle time of the current reported value. Fixes #4880.