Add two preferences (boo) for visibility:  * Hide Away Contacts (Fixes #10690). This is mostly for services like GTalk which don't do Idle properly.  * Hide Social Networking Contacts. This is for people like CFM who don't use groups and don't want Twitter contacts to show up. This fixes a strange bug in the visibility. If the user is using the default view settings (i.e., Contact Hiding plist isn't written to disk yet), when the hiding plugin registers its defaults, the hiding controller has already registered its preferences. Since the hiding controller isn't loaded as a plugin, there's no way to guarantee this order of loading.
To correct this bug, and potentially others under the same set of circumstances, I've made the preference controller inform observers of a changed key of {{{nil}}} when defaults are loaded. I am not sure of the repercussions of this, but I don't believe it will kill performance all that badly. We'll have to see, another solution (such as the hiding plugin itself informing of a {{{nil}}}-key change) might have to take place.

This fixes a situation where the menu items would say one thing, but the actual hiding behaved a different way, until a preference toggle occured.