Expose the cached user icon source, as well as the (already-defined but not public) server-side source. These two accessors let one get the NSData from the given source.
When the Twitter account is returning a serverside account, follow the same logic before, but only return the service icon if we don't have a cached image for the contact. When we do have a cached icon, return nil. This allows the context-loaded messages to have icons defined, but newly-added strangers have a temporary icon set for the same reasons as before.