(Required) OAuth support for Twitter. Added OAuthConsumer.framework as our OAuth provider. Modified MGTwitterEngine to handle OAuth, a very minor change really.
This utilizes the "save password via internalObjectID only" code I added in [26334]. A nifty side-effect of this change is that the user can change their UID all they want on Twitter.com, but it will confuse Adium for the sake of Timelines. This will be fixed later.

'''THERE IS CURRENTLY LIMITED ERROR HANDLING IF YOU DEAUTHORIZE ADIUM'''. Do not do this, you'll see a "enter password" dialog.

Remember: OAuth is in a public beta. Please report problems, but don't expect a lot.