Store the topic and topic setter in an AIChat. Add an accessor to AIAccount which determines if group chats can set topics (currently this only sets it to YES for IRC?I think XMPP supports it also, so in a future commit?). Adds a method to AIAccount to set the topic in a chat, which IRC implements in a very crude manner.
When the WKMV controller primes the view, it triggers a topic to be set for itself if the account supports it. This way, if a chat doesn't already have a topic set we set up the topic with a blank one.

The method which I copied the IRC set from is in irc.c in libpurple, but not in irc.h?irc_set_topic. Using the /command route would cause us to never be able to clear the topic.