Corrected the value of regexpUnsafeNature.
enum counts up by 1, which means that regexpUnsafeNature had value 3 (0b11), which is the value you get by setting the first two bits. In other words, regexpUnsafeNature = whitespaceNature (0b1) + shellUnsafeNature (0b10), which is bad and wrong.

The correct value is 4, which is the next higher bit (0b100).

In order to prevent mistakes like this from happening in the future, I also changed how we express the values. Previously, we used hexadecimal literals. Now we use bit-shift expressions.