Possibly correct a hang when dragging contacts around. Refs #12233.
The only explanation I can see for infinite looping in this while() is that orderIndexforObject is INFINITY, so our constant incrementing is yielding more INFINITY, ad infinitum.

I've added an AILog so that if this is truly the case it'll show up in a debug log. I'd be interested to see if this guess is right, especially since I can't figure out why it would happen except for corrupt index caches.
(transplanted from 4a64255881c8b5c3e65187efb00b531ee3ac9384)