Render the attributed string off the main thread, since large logs with lots of links (ex: log twitter timelines) can make the app spin for unacceptable lengths of time performing -[NSMutableAttributedString(AIAttributedStringAdditions) addFormattingForLinks].
Keep track of the last operation we started, then cancel it so we don't clobber a newer log, if the user has moved on.  This is better than spinning, at least, but the intensive part of the operation (adding blue/underline link styles to link attributes) is uncancellable.  So, periods of seemingly unexplained CPU use may result if a user selects another log if the current takes a while to load..