XMPP-MUC: Check log files for last activity in a channel when rejoining, to request only the necessary historic messages.
Anything that indicates the user has received messages up to then can be used (so a <status /> or <message />). It takes that tag's timestamp, and requests everything from one second after that.

This does not guarantee no double messages or shown or no messages being lost: messages are being logged with the local timestamp, and the server has its own timestamp, which can differ a lot. I think this is the best solution that doesn't require merging the histories.